requirements.txt
asgiref==3.3.4
certifi==2021.5.30
chardet==4.0.0
cloudinary==1.25.0
Django==3.2.4
django-cloudinary-storage==0.3.0
django-cors-headers==3.7.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
idna==2.10
Pillow==8.2.0
psycopg2-binary==2.8.6
PyJWT==2.1.0
python-dotenv==0.17.1
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.5
whitenoise==5.2.0
